add source selection type and flexible aspect ratio
The last missing feature is being able to select a subset of the image
as a source for copying. Apparently there is no direct way of taking a
subset out of a renderable so I might have to do a detour to javascript
and turn the canvas into an image again, then into a texture and then
use the sprite function.

Also, now we don't have a fixed aspect ratio from the start anymore,
else the custom subset descried above would be too constrained.
Therefore we keep the aspect ratio flexible until the first selection is
done, and all following selections are bound to this ratio.