fix scaling issue for changed source
I found out the transformations are applied in reverse order, so my previous attempt was doomed to fail. Now, adding the recursive images works by first translating the source back to the canvas origin, then scaling the source to the target selection size, and then moving it to the selection rectangle.
The source should contain all sink selections, else the "recursion diverges", as the same movement is repeated in every recursive step. If the movement is not "inwards" they recusive images will end up all over the place.